feat: moved authOptions to its own file in lib, moved login page to app folder, set redirects depending on whether the user is logged in or not.
