+++ src/app/(auth)/login/page.jsx
@@ -1,13 +1,11 @@
+'use client';
 import { signIn } from 'next-auth/react';
-import { getServerSession } from 'next-auth/next';
-import { authOptions } from './api/auth/[...nextauth]';
-
 import Button from '@/components/Button';
 import Image from 'next/image';
 import landingPageLogo from '@public/landingPageLogo.svg';
-import '../app/globals.css';
+import '@/app/globals.css';
 
-export default function SignIn({ providers }) {
+export default function SignIn() {
   return (
     <div className="flex h-screen flex-col items-center justify-around">
       <div className="flex flex-col items-center justify-between">
@@ -21,23 +19,10 @@ export default function SignIn({ providers }) {
         </div>
         <p className="pb-40 lg:pb-32">Technical issue report system.</p>
 
-        <Button onClick={() => signIn('42-school')}>Login</Button>
+        <Button onClick={() => signIn('42-school', { callbackUrl: '/' })}>
+          Login
+        </Button>
       </div>
     </div>
   );
-}
-
-export async function getServerSideProps(context) {
-  const session = await getServerSession(context.req, context.res, authOptions);
-
-  // If the user is already logged in, redirect.
-  // Note: Make sure not to redirect to the same page
-  // To avoid an infinite loop!
-  if (session) {
-    return { redirect: { destination: '/' } };
-  }
-
-  return {
-    props: {},
-  };
 }
