+++ src/app/api/auth/[...nextauth]/_route.js
@@ -1,0 +1,7 @@
+import NextAuth from 'next-auth';
+
+import { authOptions } from '@/lib/auth';
+
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
